Create Tree.java
This repository contains Java implementations and examples of Binary Tree data structures. It includes basic operations such as insertion,  traversal (in-order ), searching, Count Node,Count leaf and Height and more. The code is well-documented and includes comments to help understand the functionality of each method. Ideal for students, developers, or anyone interested in learning about Binary Trees and their applications in Java.