Create Doubly Linked list
This repository contains a simple yet comprehensive implementation of a Doubly Linked List in  Java. A doubly linked list is a data structure that consists of nodes where each node contains a reference to the next node and the previous node in the sequence. This allows for efficient bidirectional traversal of the list.