Create Prims_Algorithm.java
This repository contains an implementation of Prim's Algorithm in Java to find the Minimum Spanning Tree (MST) of a connected, undirected, weighted graph. Prim's Algorithm is a greedy approach that grows the MST one vertex at a time, starting from an arbitrary vertex and adding the shortest edge that connects a new vertex to the tree.