Create Heap Sort Algorithm
This repository contains an implementation of the Heap Sort algorithm in Java.

Description
Heap Sort is a comparison-based sorting algorithm that uses a binary heap data structure. It works by first building a max heap from the input data and then repeatedly extracting the maximum element from the heap and moving it to the sorted portion of the array.

Key Features:
Time Complexity: O(n log n) in all cases (best, average, and worst-case).
Space Complexity: O(1) (in-place sorting).
Stable: No, it does not maintain the relative order of equal elements.
Usage: Effective for sorting large datasets with minimal space overhead.
How It Works
Build Max Heap: Convert the input array into a max heap, where the largest element is at the root.
Extract Max Element: Swap the root of the heap with the last element of the heap and reduce the heap size by 1.
Heapify: Heapify the root element to maintain the max heap property.
Repeat the process until all elements are sorted.